arch/all-hosted: Compiler de-lint

git-svn-id: https://svn.aros.org/svn/aros/trunk@37354 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/all-hosted/bootstrap/elfloader32.c
@@ -337,7 +337,7 @@ SysBase_no:     s = sym->value;
             break;
 #endif
 	default:
-	    fprintf(stderr, "[ELF Loader] Unrecognized relocation type %d %ld\n", i, ELF_R_TYPE(rel->info));
+	    fprintf(stderr, "[ELF Loader] Unrecognized relocation type %d %ld\n", i, (long)ELF_R_TYPE(rel->info));
 	    return 0;
 	}
 	DREL(fprintf(stderr, " -> %p\n", *p));
